# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "round-1",
        "description": "round-1 doc",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/cards": {
            "get": {
                "tags": [
                    "card"
                ],
                "summary": "Get all user cards",
                "description": "Endpoint to Get all user cards",
                "operationId": "13dbd05bb9d56516aef9d29e59b54434",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cards/store": {
            "post": {
                "tags": [
                    "card"
                ],
                "summary": "Submit credit card information",
                "description": "Endpoint to submit credit card information",
                "operationId": "116f4a6197c3d7db4acabcd0b742debc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "card_name",
                                    "card_number",
                                    "card_expiry_date",
                                    "card_cvv"
                                ],
                                "properties": {
                                    "card_name": {
                                        "description": "Name on the credit card",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "card_number": {
                                        "description": "Credit card number",
                                        "type": "string",
                                        "example": "1234567812345678"
                                    },
                                    "card_expiry_date": {
                                        "description": "Expiry date of the credit card (MM/YY)",
                                        "type": "string",
                                        "example": "12/25"
                                    },
                                    "card_cvv": {
                                        "description": "CVV of the credit card",
                                        "type": "string",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cards/{id}": {
            "delete": {
                "tags": [
                    "card"
                ],
                "summary": "Delete a user card by id",
                "description": "Endpoint to delete a user card by id",
                "operationId": "9ecdb7a8f3f757d62977d82e7400ad5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "required": [
                    "status",
                    "message"
                ],
                "properties": {
                    "status": {
                        "description": "Status of the API response",
                        "type": "string"
                    },
                    "message": {
                        "description": "Message describing the result",
                        "type": "string"
                    },
                    "data": {
                        "description": "Optional data payload",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}